create execute_request.go && refactor func
