--- x/exp/keeper/execute.go
@@ -1,1 +1,0 @@
-package keeper
